refactor: move AensName to tx builder constants
