+++ src/ae/aens.ts
@@ -25,12 +25,12 @@
 
 import { salt } from '../utils/crypto'
 import { commitmentHash, isAuctionName, Pointer } from '../tx/builder/helpers'
-import { CLIENT_TTL, NAME_TTL, TX_TYPE } from '../tx/builder/schema'
+import { CLIENT_TTL, NAME_TTL, TX_TYPE, AensName } from '../tx/builder/schema'
 import { ArgumentError } from '../utils/errors'
 import { EncodedData } from '../utils/encoder'
 import { BigNumber } from 'bignumber.js'
 import { send } from './spend'
-import { AensName, getName, height } from '../chain'
+import { getName, height } from '../chain'
 import { _buildTx, BuildTxOptions } from '../tx'
 import { TransformNodeType } from '../node'
 import { NameEntry } from '../apis/node'
