+++ src/contract/aci.ts
@@ -16,7 +16,7 @@
 */
 // @ts-expect-error TODO remove
 import { Encoder as Calldata } from '@aeternity/aepp-calldata'
-import { DRY_RUN_ACCOUNT, GAS_MAX, TX_TYPE, AMOUNT } from '../tx/builder/schema'
+import { DRY_RUN_ACCOUNT, GAS_MAX, TX_TYPE, AMOUNT, AensName } from '../tx/builder/schema'
 import { buildContractIdByContractTx, unpackTx } from '../tx/builder'
 import { _buildTx } from '../tx'
 import { send } from '../ae/spend'
@@ -43,7 +43,7 @@ import { Aci as BaseAci } from '../apis/compiler'
 import Compiler from './compiler'
 import Node from '../node'
 import {
-  getAccount, getContract, getContractByteCode, getKeyBlock, resolveName, txDryRun, AensName
+  getAccount, getContract, getContractByteCode, getKeyBlock, resolveName, txDryRun
 } from '../chain'
 import AccountBase from '../account/base'
